// src/apis/user/user.enum.ts
export enum UserType {
  ROOT = 'Root',
  ADMIN = 'Admin',
  HR = 'HR',
  User = 'User',
  CANDIDATE = 'Candidate',
  EMPLOYEE = 'Employee', // ứng viên chấp nhận vào làm việc thì sẽ chuyển từ Candidate sang Employee
  C_B = 'C&B',
}
export enum UserStatus {
  ACTIVE = 'Đang làm việc',
  INACTIVE = 'Không hoạt động',
}

export enum Gender {
  MALE = 'Nam',
  FEMALE = 'Nữ',
  OTHER = 'Khác',
}

export enum UserMovementType {
  APPOINTMENT = 'Bổ nhiệm', // Bổ nhiệm
  TRANSFER = 'Điều chuyển', // Điều chuyển
  DEMOTION = 'Miễn nhiệm', // Miễn nhiệm
}

export enum UserMovementStatus {
  PENDING = 'Đang chờ',
  APPROVED = 'Đã phê duyệt',
  REJECTED = 'Đã từ chối',
  EFFECTIVE = 'Có hiệu lực',
  CANCELLED = 'Hủy ban hành',
  VIEW = 'Xem',
}

export interface FlatArray {
  id: string;
  name: string;
  status: string;
}

export enum UserMovementSearchType {
  MY_PROPOSAL = 'Đề xuất của tôi',
  NEED_APPROVE = 'Đề xuất cần phê duyệt',
  ALL = 'Tất cả',
}

export enum UserMovementHistoryField {
  REASON = 'reason',
  DATE_APPOINTMENT = 'dateAppointment',
  FILE = 'file',
}

export enum UserOfficialStatus {
  OFFICIAL = 'Chính thức',
  TRIAL = 'Thử việc',
}